Switch to a worker-less approach
