--- .github/workflows/wrangler.yaml
@@ -1,22 +1,0 @@
-name: Deploy packages.element.io Wrangler
-on:
-  push:
-    branches: [ develop ]
-    paths:
-      - 'packages.element.io/**'
-concurrency:
-  group: ${{ github.workflow }}-${{ github.ref }}
-  cancel-in-progress: true
-jobs:
-  wrangler:
-    name: "Deploy"
-    runs-on: ubuntu-latest
-    environment: develop
-    steps:
-      - uses: actions/checkout@v2
-
-      - name: Publish
-        uses: cloudflare/wrangler-action@2.0.0
-        with:
-          apiToken: ${{ secrets.CF_API_TOKEN }}
-          workingDirectory: 'packages.element.io'
