+++ package.json
@@ -50,6 +50,7 @@
     "request": "^2.88.2"
   },
   "devDependencies": {
+    "@aws-sdk/client-s3": "^3.213.0",
     "@babel/core": "^7.18.10",
     "@babel/preset-env": "^7.18.10",
     "@babel/preset-typescript": "^7.18.6",
@@ -93,7 +94,7 @@
     "rimraf": "^3.0.2",
     "tar": "^6.1.2",
     "ts-jest": "^28.0.8",
-    "ts-node": "^10.4.0",
+    "ts-node": "^10.9.1",
     "typescript": "4.5.5"
   },
   "hakDependencies": {
