+++ packages.element.io/fonts/Inter-BoldItalic.woff
