+++ packages.element.io/fonts/Inter-ExtraBoldItalic.woff2
