+++ packages.element.io/fonts/Inter-italic.var.woff2
