--- packages.element.io/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "packages-element-io",
-  "version": "0.0.0",
-  "devDependencies": {
-    "@cloudflare/kv-asset-handler": "^0.2.0",
-    "@cloudflare/workers-types": "^3.18.0",
-    "@worker-tools/html": "^2.0.0-pre.13",
-    "typescript": "^4.9.3",
-    "wrangler": "2.3.1"
-  },
-  "private": true,
-  "scripts": {
-    "start": "wrangler dev",
-    "deploy": "wrangler publish"
-  }
-}
