--- packages.element.io/wrangler.toml
@@ -1,11 +1,0 @@
-name = "packages-element-io"
-main = "src/index.ts"
-compatibility_date = "2022-11-16"
-
-[site]
-bucket = "./public"
-
-[[r2_buckets]]
-binding = "PACKAGES_BUCKET"
-bucket_name = "packages-element-io"
-preview_bucket_name = "packages-element-io"
